# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Get information about countries via a RESTful API made by Go.\nSource Code: https://github.com/efectn/rest-countries",
        "title": "Rest Countries Go",
        "contact": {},
        "license": {
            "name": "GPL-3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/all": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Get all countries.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/calling/{code}": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by calling code.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Calling Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/capital/{capital}": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by capital city.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Capital city",
                        "name": "capital",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/code/{code}": {
            "get": {
                "description": "Single (tr), Plural (tr;gb;usa).",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by ISO 3166-1 code.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISO 3166-1 Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/currency/{currency}": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by ISO 4217 currency code.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISO 4217 Currency Code",
                        "name": "currency",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/language/{language}": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by ISO 639-1 language code.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISO 639-1 Language Code",
                        "name": "language",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/name/{name}": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by name or native name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name, Native Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Fulltext Search",
                        "name": "fullText",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/region/{region}": {
            "get": {
                "description": "Regions: Africa, Americas, Asia, Europe, Oceania.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by region.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/regional-block/{block}": {
            "get": {
                "description": "Blocks:\n- EU (European Union)\n- EFTA (European Free Trade Association)\n- CARICOM (Caribbean Community)\n- PA (Pacific Alliance)\n- AU (African Union)\n- USAN (Union of South American Nations)\n- EEU (Eurasian Economic Union)\n- AL (Arab League)\n- ASEAN (Association of Southeast Asian Nations)\n- CAIS (Central American Integration System)\n- CEFTA (Central European Free Trade Agreement)\n- NAFTA (North American Free Trade Agreement)\n- SAARC (South Asian Association for Regional Cooperation)",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main"
                ],
                "summary": "Search countries by regional block.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Regional Block",
                        "name": "block",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}